captcha-recognize/requirements.txt
requests==2.18.4
Pillow==4.2.1
